requirements.txt
fastapi
uvicorn[standard]
anthropic
chromadb
python-dotenv
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
yt-dlp
rapidfuzz
